% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{createBarPrior}
\alias{createBarPrior}
\title{Create a BAR Cyclops prior object}
\usage{
createBarPrior(penalty = "bic", exclude = c(), forceIntercept = FALSE)
}
\arguments{
\item{penalty}{Specifies the BAR penalty; possible values are `BIC` or `AIC` or a numeric value}

\item{exclude}{A vector of numbers or covariateId names to exclude from prior}

\item{forceIntercept}{Logical: Force intercept coefficient into prior}
}
\value{
A BAR Cyclops prior object of class inheriting from
\code{"cyclopsPrior"} for use with \code{fitCyclopsModel}.
}
\description{
\code{createBarPrior} creates a BAR Cyclops prior object for use with \code{\link{fitCyclopsModel}}.
}
\examples{
prior <- createBarPrior(penalty = "bic")

}
